Change proto of printi

+++ 0x02-functions_nested_loops/holberton.h
@@ -15,7 +15,7 @@ int add(int, int);
 void print_to_98(int n);
 void print_times_table(int n);
 void print_space(int x);
-void printi(int n);
+void printi(unsigned long n);
 
 #endif
